Added password-recovery mail, and security things with by 401 & 403
--
Daniel Ripoll <info@danielripoll.es>
http://danielripoll.es
+34 6688 27 4 90
LSYS Network http://lsys.net

+++ routes/login.rb
@@ -1,27 +1,36 @@
 class ASYD < Sinatra::Application
   get '/login' do
-    status 200
     if !File.directory? 'data'
       redirect '/setup'
     end
-    erb :login, :layout => false
+    if session[:username].blank?
+      erb :login, :layout => false
+    else
+      redirect '/'
+    end
   end
 
   post '/login' do
-    status 200
-    u = User.auth(params['username'], params['password'])
-    if u
-      session[:username] = u.username
-      redirect '/'
+    if session[:username].blank?
+      u = User.auth(params['username'], params['password'])
+      if u
+        session[:username] = u.username
+        redirect '/'
+      else
+        @error = "Failed login"
+        erb :login, :layout => false
+      end
     else
-      @error = "Failed login"
-      erb :login, :layout => false
+      redirect '/'
     end
   end
 
   get '/logout' do
-    status 200
-    session.delete(:username)
-    erb "<div class='alert alert-message'>Logged out</div>"
+    if !session[:username].blank?
+      session.delete(:username)
+      erb "<div class='alert alert-message'>Logged out</div>"
+    else
+      redirect '/login'
+    end
   end
 end
